fix(review): fala 2 — preview fidelity, tofu .notdef, int64 freq penalty
- gui: smart preview syncs allow_mirror/edge_aware from checkboxes
- gui: typo preview uses group-filtered engine (cached by group set)
- indexer_typo: skip codepoints not in font cmap (eliminates .notdef tofu)
- engine_smart: used_counts int64 to avoid **2 overflow on huge renders

+++ src/engine_smart.py
@@ -646,7 +646,10 @@ class SmartEngine:
             print("  Edge-Aware Matching active (79-dim features)")
 
         tgt_features = np.array([x["feature"] for x in sectors_data])
-        used_counts = np.zeros(len(self.paths), dtype=np.int32)
+        # int64: used_counts**2 in the frequency penalty (below) overflows int32
+        # once a tile is reused >46340 times (huge render + tiny library), which
+        # would wrap negative and invert the penalty.
+        used_counts = np.zeros(len(self.paths), dtype=np.int64)
         sector_assignments = -1 * np.ones(len(sectors_data), dtype=np.int32)
         failed_tiles = 0
